The Editor looks nicer, and selection works.
The editor clearly isn't as good as it could be, but it's more user friendly than it was.
Notably missing functionality: You can't actually edit an object or structure which has already been created (Currently, the only option is to delete the object and recreate it with new properties)
The top priority now is variable map size.