Can now undo while playing.
State changes are put on a stack in the form of Deltas.  It hasn't been completely implemented - ordinary motion and sticky block groups can be undone, but switch mechanisms don't behave well.  The next goal is to make undo work for switches as well, and also fix the way that Structures are stored.